+++ src/subcommands/find.rs
@@ -1,6 +1,6 @@
 use rust_releases::{Release, ReleaseIndex};
 
-use crate::check::RunCheck;
+use crate::check::Check;
 use crate::config::{Config, ModeIntent, SearchMethod};
 use crate::errors::{CargoMSRVError, TResult};
 use crate::releases::filter_releases;
@@ -8,13 +8,35 @@ use crate::reporter::Output;
 use crate::result::MinimalCompatibility;
 use crate::search_methods::{Bisect, FindMinimalCapableToolchain, Linear};
 use crate::toolchain_file::write_toolchain_file;
+use crate::SubCommand;
 
-pub fn run_find_msrv_action<R: Output>(
+pub struct Find<'index, C: Check> {
+    release_index: &'index ReleaseIndex,
+    runner: C,
+}
+
+impl<'index, C: Check> Find<'index, C> {
+    pub fn new(release_index: &'index ReleaseIndex, runner: C) -> Self {
+        Self {
+            release_index,
+            runner,
+        }
+    }
+}
+
+impl<'index, C: Check> SubCommand for Find<'index, C> {
+    fn run<R: Output>(&self, config: &Config, reporter: &R) -> TResult<()> {
+        find_msrv(config, reporter, self.release_index, &self.runner)
+    }
+}
+
+fn find_msrv<R: Output, C: Check>(
     config: &Config,
     reporter: &R,
     release_index: &ReleaseIndex,
+    runner: &C,
 ) -> TResult<()> {
-    match find_msrv(config, reporter, release_index)? {
+    match search(config, reporter, release_index, runner)? {
         MinimalCompatibility::NoCompatibleToolchains => {
             info!("no minimal-compatible toolchain found");
 
@@ -37,51 +59,52 @@ pub fn run_find_msrv_action<R: Output>(
     }
 }
 
-pub fn find_msrv<R: Output>(
+fn search<R: Output, C: Check>(
     config: &Config,
     reporter: &R,
     index: &rust_releases::ReleaseIndex,
+    runner: &C,
 ) -> TResult<MinimalCompatibility> {
     let releases = index.releases();
     let included_releases = filter_releases(config, releases);
 
     reporter.mode(ModeIntent::Find);
     reporter.set_steps(included_releases.len() as u64);
-    run_with_search_method(config, &included_releases, reporter)
+
+    run_with_search_method(config, &included_releases, reporter, runner)
 }
 
-fn run_with_search_method(
+fn run_with_search_method<R: Output, C: Check>(
     config: &Config,
     included_releases: &[Release],
-    output: &impl Output,
+    reporter: &R,
+    runner: &C,
 ) -> TResult<MinimalCompatibility> {
-    output.set_steps(included_releases.len() as u64);
+    reporter.set_steps(included_releases.len() as u64);
 
     let search_method = config.search_method();
     info!(?search_method);
 
-    let runner = RunCheck::new(output);
-
     // Run a linear or binary search depending on the configuration
     match search_method {
         SearchMethod::Linear => {
-            run_searcher(&Linear::new(runner), included_releases, config, output)
+            run_searcher(&Linear::new(runner), included_releases, config, reporter)
         }
         SearchMethod::Bisect => {
-            run_searcher(&Bisect::new(runner), included_releases, config, output)
+            run_searcher(&Bisect::new(runner), included_releases, config, reporter)
         }
     }
 }
 
-fn run_searcher(
+fn run_searcher<R: Output>(
     method: &impl FindMinimalCapableToolchain,
     releases: &[Release],
     config: &Config,
-    output: &impl Output,
+    reporter: &R,
 ) -> TResult<MinimalCompatibility> {
-    let minimum_capable = method.find_toolchain(releases, config, output)?;
+    let minimum_capable = method.find_toolchain(releases, config, reporter)?;
 
-    report_outcome(&minimum_capable, config, output);
+    report_outcome(&minimum_capable, config, reporter);
 
     Ok(minimum_capable)
 }
@@ -96,3 +119,6 @@ fn report_outcome(minimum_capable: &MinimalCompatibility, config: &Config, outpu
         }
     }
 }
+
+#[cfg(test)]
+mod tests;